refactor: 수식과 관련된 기능 Expression 클래스로 분리
